requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-crispy-forms==2.2
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.7
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
